requirements.txt
flask
requests
Pillow
gunicorn